Add new shape.
