+++ crates/types/src/enums.rs
@@ -1,11 +1,9 @@
 use crate::literals::LiteralValue;
-use crate::SchemaField;
+use crate::schema::SchemaField;
 
 #[derive(Clone, Debug, Default)]
 pub struct EnumType {
     pub default_index: Option<usize>,
-    pub description: Option<String>,
-    pub name: Option<String>,
     pub values: Vec<LiteralValue>,
     pub variants: Option<Vec<SchemaField>>,
 }
