+++ crates/types/src/objects.rs
@@ -6,7 +6,6 @@ pub struct ObjectType {
     pub key_type: Box<SchemaType>,
     pub max_length: Option<usize>,
     pub min_length: Option<usize>,
-    pub name: Option<String>,
     pub required: Option<Vec<String>>,
     pub value_type: Box<SchemaType>,
 }
