+++ crates/types/src/strings.rs
@@ -10,7 +10,6 @@ pub struct StringType {
     pub format: Option<String>,
     pub max_length: Option<usize>,
     pub min_length: Option<usize>,
-    pub name: Option<String>,
     pub pattern: Option<String>,
 }
