+++ crates/types/src/structs.rs
@@ -1,10 +1,8 @@
-use crate::SchemaField;
+use crate::schema::SchemaField;
 
 #[derive(Clone, Debug, Default)]
 pub struct StructType {
-    pub description: Option<String>,
     pub fields: Vec<SchemaField>,
-    pub name: Option<String>,
     pub partial: bool,
     pub required: Option<Vec<String>>,
 }
